Creating Find_Max_Hourglass_Number
Given a 6x6 array, I define an hourglass in array to be a subset of values with indices falling in this pattern in array's graphical representation: 
a b c
   d   
e f g
There are 16 hourglasses in array, and an hourglass sum is the sum of an hourglass' values. 
The task is calculating the hourglass sum for every hourglass in array, then printing the maximum hourglass sum.